refactor: move stuff out of main
