+++ CMakeLists.txt
@@ -126,9 +126,9 @@ endif()
 
 # Add executable with platform-specific source files
 if(WIN32)
-    add_executable(MyCMakeCrasher main.cpp windows.cpp)
+    add_executable(MyCMakeCrasher main.cpp paths.cpp windows.cpp)
 else()
-    add_executable(MyCMakeCrasher main.cpp)
+    add_executable(MyCMakeCrasher main.cpp paths.cpp)
 endif()
 
 # Platform-specific libraries and link order
